+++ CHANGELOG.md
@@ -8,6 +8,25 @@ If you like this project and find it useful, please consider giving it a star on
   <img src="bmc-button.svg" alt="Buy me a coffee" width="120">
 </a>
 
+## [3.2.5] - 2025-09-01
+
+### Added
+
+- [refrigerator]: Added RefrigeratorAlarm cluster.
+
+### Changed
+
+- [package]: Updated dependencies.
+- [devContainer]: Updated devContainer with repository name for the container.
+
+### Fixed
+
+- [refrigerator]: Fix device type
+
+<a href="https://www.buymeacoffee.com/luligugithub">
+  <img src="bmc-button.svg" alt="Buy me a coffee" width="80">
+</a>
+
 ## [3.2.4] - 2025-08-29
 
 ### Added
@@ -17,8 +36,6 @@ If you like this project and find it useful, please consider giving it a star on
 ### Changed
 
 - [package]: Updated dependencies.
-
-### Fixed
 
 <a href="https://www.buymeacoffee.com/luligugithub">
   <img src="bmc-button.svg" alt="Buy me a coffee" width="80">